Update test model packages
