+++ requery-test/src/main/java/io/requery/test/jpa/AddressType.java
@@ -14,23 +14,9 @@
  * limitations under the License.
  */
 
-package io.requery.test.modeljpa;
+package io.requery.test.jpa;
 
-import javax.persistence.Entity;
-import javax.persistence.GeneratedValue;
-import javax.persistence.Id;
-import javax.persistence.ManyToOne;
-import java.io.Serializable;
-
-@Entity
-public interface Phone extends Serializable {
-
-    @Id
-    @GeneratedValue
-    int getId();
-    String getPhoneNumber();
-    boolean isNormalized();
-
-    @ManyToOne
-    Person getOwner();
+public enum AddressType {
+    HOME,
+    WORK
 }
