+++ requery-test/src/main/java/io/requery/test/jpa/Group.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package io.requery.test.modeljpa;
+package io.requery.test.jpa;
 
 
 import io.requery.Persistable;
